--- src/containers/shared/components/Transaction/default/parser.ts
@@ -1,1 +1,0 @@
-export const parser = (tx: any): any => tx
