+++ src/containers/shared/components/Transaction/defaultParser.ts
@@ -1,0 +1,1 @@
+export const defaultParser = (tx: any): any => tx
